Reorganize reducers and state management to be more efficient
