+++ server/ducks/occurrences.js
@@ -1,14 +1,10 @@
-/* eslint new-cap:0, no-eval:0 */
-import { lightState } from 'node-hue-api';
-import { mapValues, get } from 'lodash';
+/* eslint no-eval:0 */
+import { get } from 'lodash';
 
 import { config } from 'environment';
-import { EVENT_EMITTERS_MAP, CUSTOM_LIGHT_FUNCTIONS } from 'constants';
+import { CUSTOM_LIGHT_FUNCTIONS } from 'constants';
 import {
-  initCustomState,
-  handleAction,
   toggleLight,
-  logActionType,
   logUndefinedHandler,
   errorLightStatus
 } from 'utils';
@@ -19,68 +15,50 @@ export const EMIT_FORWARD_HTTP_REQUEST = 'EMIT_FORWARD_HTTP_REQUEST';
 export const EMIT_SEND_UNIFIED_COMMAND = 'EMIT_SEND_UNIFIED_COMMAND';
 export const EMIT_CUSTOM_STATE_UPDATE = 'EMIT_CUSTOM_STATE_UPDATE';
 
-const initialState = {
-  lightState,
-  ...mapValues(initCustomState(config), 'default')
-};
+const occurrencesReducer = (state, action) => ({
+  [EMIT_SEND_HUE_COMMAND]() {
+    const { hue } = state;
+    const ipaddress = action.ipaddress || Object.keys(hue.userIDs)[0];
+    const hueBridge = get(hue, `['${ipaddress}'].bridge`);
 
-const occurrencesReducer = (state = initialState, action) => {
-  logActionType(action.type);
-
-  const reducers = {
-    [EMIT_SEND_HUE_COMMAND]() {
-      const { hue } = action.devicesReducer;
-      const ipaddress = action.ipaddress || Object.keys(hue.userIDs)[0];
-      const hueBridge = get(hue, `['${ipaddress}'].bridge`);
-
-      if (CUSTOM_LIGHT_FUNCTIONS.includes(action.func)) {
-        switch (action.func) {
-          case 'toggle':
-            toggleLight(hueBridge, state, action.id);
-            break;
-          default:
-            errorLightStatus();
-            break;
-        }
-      } else {
-        try {
-          const newLightState = state.lightState.create()[action.func](action.arg);
-
-          hueBridge.setLightState(action.id, newLightState);
-        } catch (e) {
+    if (CUSTOM_LIGHT_FUNCTIONS.includes(action.func)) {
+      switch (action.func) {
+        case 'toggle':
+          toggleLight(hueBridge, state, action.id);
+          break;
+        default:
           errorLightStatus();
-        }
+          break;
       }
+    } else {
+      try {
+        const newLightState = state.lightState.create()[action.func](action.arg);
 
-      return { ...state };
-    },
-
-    [EMIT_CUSTOM_STATE_UPDATE](customStateConfig) {
-      action.$state.forEach((customStateKey) => {
-        try {
-          const forcedHandler = get(config, `${action.$path}.$state[${customStateKey}].$handler`);
-          const riderHandler = () => customStateConfig[customStateKey].$handler;
-
-          const customStateHandler = eval(forcedHandler || riderHandler());
-
-          state[customStateKey] = customStateHandler(state[customStateKey]);
-        } catch (e) {
-          logUndefinedHandler(e);
-        }
-      });
-
-      return { ...state };
+        hueBridge.setLightState(action.id, newLightState);
+      } catch (e) {
+        errorLightStatus();
+      }
     }
-  };
 
-  const eventKey = EVENT_EMITTERS_MAP[action.type];
-  const customStateConfig = get(config[eventKey], `${action.key}.$state`);
+    return { ...state };
+  },
 
-  if (customStateConfig && action.$state) {
-    reducers[EMIT_CUSTOM_STATE_UPDATE](customStateConfig);
+  [EMIT_CUSTOM_STATE_UPDATE](customStateConfig) {
+    action.$state.forEach((customStateKey) => {
+      try {
+        const forcedHandler = get(config, `${action.$path}.$state[${customStateKey}].$handler`);
+        const riderHandler = () => customStateConfig[customStateKey].$handler;
+
+        const customStateHandler = eval(forcedHandler || riderHandler());
+
+        state[customStateKey] = customStateHandler(state[customStateKey]);
+      } catch (e) {
+        logUndefinedHandler(e);
+      }
+    });
+
+    return { ...state };
   }
-
-  return handleAction(state, action, reducers);
-};
+});
 
 export default occurrencesReducer;
